Change ImagesController implementation to store image files not only in db

+++ PropertyRental.Application/Images/Commands/DeleteImage/DeleteImageCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using MediatR;
 using Microsoft.EntityFrameworkCore;
+using PropertyRental.Application.Common.Exceptions;
 using PropertyRental.Application.Common.Interfaces;
 using System.Linq;
 using System.Threading;
@@ -19,6 +20,9 @@ namespace PropertyRental.Application.Images.Commands.DeleteImage
 		public async Task<Unit> Handle(DeleteImageCommand request, CancellationToken cancellationToken)
 		{
 			var image = await _context.Images.Where(i => i.Id == request.ImageId).FirstOrDefaultAsync(cancellationToken);
+
+			if (image == null)
+				throw new CommandException("Can not delete: Image does not exists");
 
 			_context.Images.Remove(image);
